Break apart main entrypoint into smaller modules
